Fixed bug with collecting items and reimported assets(because Unity was updated)

+++ Assets/Scripts/Controllers/GameManager.cs
@@ -7,7 +7,7 @@ public class GameManager : MonoBehaviour
 {
     public static GameManager instance;
     
-    public int score;       
+    public double score;       
     public int lifescore;
 
     public bool HeroDiedAndRestart;
